refactor(analysis): split utilityStats and group tabs into folders
The `analysis/` folder had grown to 24 flat files, and `utilityStats.ts`
had become a mis-named hub: it mixed generic round/team helpers (imported
by 16 files, several of them outside `analysis/` in `player/`) with the
utility-specific flash/damage aggregation it was named for.

- Extract the generic helpers into `domain/rounds.ts` (roundSides,
  isKnifeRound, isPreGameRound, preGameRoundCount, roundDisplayLabels,
  flashSetupForKill) and `domain/teams.ts` (Team, TeamPlayer, groupTeams).
  They are pure logic over `Replay` and shared by the player, so they
  belong in `domain/`, not under any single feature.
- `utilityStats.ts` now holds only the flash/damage stats it is named for.
- Group the tabs into folders that mirror the top-level nav:
  economy/ utility/ duels/ heatmap/, plus shared/ for the cross-tab parts
  (HeatmapPlot, RoundStrip, RoundTimeRange, roundTime, heatmapTypes).
- Update all imports and the viewer/README.md architecture map.

Pure move + re-export split, no behavior change. type-check passes.
