+++ apps/app/src/viewer/analysis/duels/DuelsView.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import type { Replay } from '@/viewer/domain/schema'
-import DuelStatsView from '@/viewer/analysis/DuelStatsView.vue'
-import OpeningDuelMapView from '@/viewer/analysis/OpeningDuelMapView.vue'
-import HeatmapView from '@/viewer/analysis/HeatmapView.vue'
+import DuelStatsView from '@/viewer/analysis/duels/DuelStatsView.vue'
+import OpeningDuelMapView from '@/viewer/analysis/duels/OpeningDuelMapView.vue'
+import HeatmapView from '@/viewer/analysis/heatmap/HeatmapView.vue'
 import { useI18n } from '@/i18n'
 
 const { t } = useI18n()
