+++ apps/app/src/viewer/analysis/duels/OpeningDuelMapView.vue
@@ -4,11 +4,11 @@ import type { Replay, Round, Side } from '@/viewer/domain/schema'
 import { MAP_CALIBRATION } from '@/viewer/domain/calibration'
 import { SIDE_COLOR } from '@/viewer/domain/colors'
 import { killWeaponIcon } from '@/viewer/domain/weaponIcons'
-import HeatmapPlot from '@/viewer/analysis/HeatmapPlot.vue'
+import HeatmapPlot from '@/viewer/analysis/shared/HeatmapPlot.vue'
 import KillfeedRow from '@/viewer/player/KillfeedRow.vue'
-import type { KillInfo } from '@/viewer/analysis/heatmapTypes'
-import { computeOpeningDuels } from '@/viewer/analysis/duelStats'
-import { groupTeams } from '@/viewer/analysis/utilityStats'
+import type { KillInfo } from '@/viewer/analysis/shared/heatmapTypes'
+import { computeOpeningDuels } from '@/viewer/analysis/duels/duelStats'
+import { groupTeams } from '@/viewer/domain/teams'
 import { useI18n } from '@/i18n'
 
 /**
