+++ apps/app/src/viewer/analysis/duels/OpeningDuelsView.vue
@@ -2,8 +2,8 @@
 import { computed, ref } from 'vue'
 import { PopoverClose, PopoverContent, PopoverPortal, PopoverRoot, PopoverTrigger } from 'reka-ui'
 import type { Replay, Side } from '@/viewer/domain/schema'
-import { groupTeams } from '@/viewer/analysis/utilityStats'
-import { computeOpeningDuels, type OpeningDuel } from '@/viewer/analysis/duelStats'
+import { groupTeams } from '@/viewer/domain/teams'
+import { computeOpeningDuels, type OpeningDuel } from '@/viewer/analysis/duels/duelStats'
 import { SIDE_COLOR } from '@/viewer/domain/colors'
 import KillfeedRow from '@/viewer/player/KillfeedRow.vue'
 import UiSwitch from '@/ui/UiSwitch.vue'
