+++ apps/app/src/viewer/analysis/heatmap/HeatmapView.vue
@@ -3,16 +3,16 @@ import { computed, ref, watch } from 'vue'
 import type { Replay, Round, Side } from '@/viewer/domain/schema'
 import { MAP_CALIBRATION } from '@/viewer/domain/calibration'
 import { SIDE_COLOR } from '@/viewer/domain/colors'
-import HeatmapPlot from '@/viewer/analysis/HeatmapPlot.vue'
+import HeatmapPlot from '@/viewer/analysis/shared/HeatmapPlot.vue'
 import KillfeedRow from '@/viewer/player/KillfeedRow.vue'
-import UtilityHeatmapView from '@/viewer/analysis/UtilityHeatmapView.vue'
-import type { KillInfo } from '@/viewer/analysis/heatmapTypes'
+import UtilityHeatmapView from '@/viewer/analysis/utility/UtilityHeatmapView.vue'
+import type { KillInfo } from '@/viewer/analysis/shared/heatmapTypes'
 import { killWeaponIcon } from '@/viewer/domain/weaponIcons'
-import { isKnifeRound, roundSides } from '@/viewer/analysis/utilityStats'
+import { isKnifeRound, roundSides } from '@/viewer/domain/rounds'
 import UiSelect from '@/ui/UiSelect.vue'
-import RoundTimeRange from '@/viewer/analysis/RoundTimeRange.vue'
-import RoundStrip from '@/viewer/analysis/RoundStrip.vue'
-import { freezeSeconds, maxLiveRoundTime } from '@/viewer/analysis/roundTime'
+import RoundTimeRange from '@/viewer/analysis/shared/RoundTimeRange.vue'
+import RoundStrip from '@/viewer/analysis/shared/RoundStrip.vue'
+import { freezeSeconds, maxLiveRoundTime } from '@/viewer/analysis/shared/roundTime'
 import { useI18n } from '@/i18n'
 
 const { t } = useI18n()
