+++ apps/app/src/viewer/analysis/shared/HeatmapPlot.vue
@@ -3,7 +3,7 @@ import { computed, onBeforeUnmount, onMounted, ref, watch } from 'vue'
 import { type MapCalibration, worldToFraction } from '@/viewer/domain/calibration'
 import type { Replay, Side } from '@/viewer/domain/schema'
 import { SIDE_COLOR } from '@/viewer/domain/colors'
-import type { KillInfo } from '@/viewer/analysis/heatmapTypes'
+import type { KillInfo } from '@/viewer/analysis/shared/heatmapTypes'
 import ReplayClipPopover from '@/viewer/player/ReplayClipPopover.vue'
 import UiIcon from '@/ui/UiIcon.vue'
 import { useI18n } from '@/i18n'
