+++ apps/app/src/viewer/analysis/utility/UtilitiesView.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import type { PlayerMeta, Replay } from '@/viewer/domain/schema'
-import UtilityThrowsView from '@/viewer/analysis/UtilityThrowsView.vue'
-import UtilityFlashesView from '@/viewer/analysis/UtilityFlashesView.vue'
-import UtilityDamageView from '@/viewer/analysis/UtilityDamageView.vue'
+import UtilityThrowsView from '@/viewer/analysis/utility/UtilityThrowsView.vue'
+import UtilityFlashesView from '@/viewer/analysis/utility/UtilityFlashesView.vue'
+import UtilityDamageView from '@/viewer/analysis/utility/UtilityDamageView.vue'
 import { useI18n } from '@/i18n'
 
 const { t } = useI18n()
